+++ include/navkit/app_support/EmulatorBindingPolicy.hpp
@@ -4,17 +4,23 @@
 #pragma once
 
 #include "navkit/app_support/EmulatorBinding.hpp"
-#include "navkit/app_support/EmulatorBindingTraits.hpp"
+#include "navkit/app_support/EmulatorPolicy.hpp"
+#include "navkit/core/estimation/sensor/SensorPolicy.hpp"
+#include "navkit/io/RunLogger.hpp"
 
 #include <concepts>
-#include <tuple>
 
 namespace navkit::app_support
 {
 
-template<typename BindingTuple, typename SensorTuple>
-concept EmulatorBindingPolicy = requires { typename std::tuple_size<BindingTuple>::type; } &&
-                                emulator_binding_ids_unique_v<BindingTuple> &&
-                                emulator_binding_sensors_valid_v<BindingTuple, SensorTuple>;
+template<typename Candidate>
+concept EmulatorBindingPolicy =
+    requires {
+        { Candidate::Id } -> std::convertible_to<SensorId>;
+        typename Candidate::Emulator_t;
+        typename Candidate::Sensor_t;
+    } && navkit::core::estimation::SensorPolicy<typename Candidate::Sensor_t> &&
+    EmulatorPolicy<typename Candidate::Emulator_t, typename Candidate::Sensor_t, io::RunLogger> &&
+    (Candidate::Id == Candidate::Emulator_t::Id) && (Candidate::Id == Candidate::Sensor_t::Id);
 
 } // namespace navkit::app_support