+++ include/navkit/core/models/GnssPosModel.hpp
@@ -5,6 +5,7 @@
 
 #include "navkit/core/estimation/measurement/MeasurementModelBase.hpp"
 #include "navkit/core/estimation/state/Segment.hpp"
+#include "navkit/core/estimation/state/StateDefPolicy.hpp"
 
 #include <string_view>
 
@@ -14,7 +15,7 @@ namespace navkit::core::models
 using navkit::core::estimation::MeasurementModelBase;
 using navkit::core::estimation::segment;
 
-template<typename StateDef>
+template<navkit::core::estimation::StateDefPolicy StateDef>
 class GnssPosModel : public MeasurementModelBase<GnssPosModel<StateDef>, StateDef, 3>
 {
 public: